Move implementation of isotherm to cc file.
